refactor: extract analytics constants
